requirements.txt
request
python-dotenv